Scene Drawer, Pickle
Changed the SceneDrawer class so that generating images and generating the pickle file can be done in separate executions. Generating the pickle file is non-deterministic because model positions are chosen randomly. So, if you make a new pickle file, you have to make the images again, too.

Uploaded the updated scene list pickle file.